// frontend/src/api/generated/models/GreetingPage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Template App API
 * API for the Template Application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PageMeta } from './PageMeta';
import {
    PageMetaFromJSON,
    PageMetaFromJSONTyped,
    PageMetaToJSON,
    PageMetaToJSONTyped,
} from './PageMeta';
import type { GreetingResponse } from './GreetingResponse';
import {
    GreetingResponseFromJSON,
    GreetingResponseFromJSONTyped,
    GreetingResponseToJSON,
    GreetingResponseToJSONTyped,
} from './GreetingResponse';

/**
 * 
 * @export
 * @interface GreetingPage
 */
export interface GreetingPage {
    /**
     * 
     * @type {Array<GreetingResponse>}
     * @memberof GreetingPage
     */
    data: Array<GreetingResponse>;
    /**
     * 
     * @type {PageMeta}
     * @memberof GreetingPage
     */
    meta: PageMeta;
}

/**
 * Check if a given object implements the GreetingPage interface.
 */
export function instanceOfGreetingPage(value: object): value is GreetingPage {
    if (!('data' in value) || value['data'] === undefined) return false;
    if (!('meta' in value) || value['meta'] === undefined) return false;
    return true;
}

export function GreetingPageFromJSON(json: any): GreetingPage {
    return GreetingPageFromJSONTyped(json, false);
}

export function GreetingPageFromJSONTyped(json: any, ignoreDiscriminator: boolean): GreetingPage {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(GreetingResponseFromJSON)),
        'meta': PageMetaFromJSON(json['meta']),
    };
}

export function GreetingPageToJSON(json: any): GreetingPage {
    return GreetingPageToJSONTyped(json, false);
}

export function GreetingPageToJSONTyped(value?: GreetingPage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(GreetingResponseToJSON)),
        'meta': PageMetaToJSON(value['meta']),
    };
}